+++ vscode/sampleprj/packages/settings/.prettierrc
@@ -1,0 +1,5 @@
+{
+    "arrowParens": "avoid",
+    "semi": false,
+    "tabWidth": 4
+}
