+++ vscode/sampleprj/packages/settings/README.md
@@ -1,0 +1,6 @@
+# @devicescript/settings
+
+A [DeviceScript](https://microsoft.github.io/devicescript/) library
+to manage settings.
+
+-   [Read documentation](https://microsoft.github.io/devicescript/developer/settings)
