+++ vscode/sampleprj/packages/settings/src/client.ts
@@ -1,0 +1,6 @@
+import * as ds from "@devicescript/core"
+
+/**
+ * The cloud adapter client
+ */
+export const settings = new ds.Settings()
