+++ vscode/sampleprj/packages/settings/src/tsconfig.json
@@ -1,0 +1,24 @@
+{
+    "compilerOptions": {
+        "moduleResolution": "node",
+        "target": "es2022",
+        "module": "es2022",
+        "lib": [],
+        "strict": true,
+        "strictNullChecks": false,
+        "strictFunctionTypes": true,
+        "sourceMap": false,
+        "declaration": false,
+        "experimentalDecorators": true,
+        "preserveConstEnums": true,
+        "noImplicitThis": true,
+        "isolatedModules": true,
+        "noImplicitAny": true,
+        "moduleDetection": "force",
+        "types": []
+    },
+    "include": [
+        "*.ts",
+        "../.devicescript/lib/*.ts"
+    ]
+}
